Updated math and rand libraries to use double.

+++ duck/lib/dmath.c
@@ -15,8 +15,8 @@ TODO:
 
 /*Math.sin(angle)*/
 int MathSin(int arg_count){
-  float angle = 0.0;
-  float val = 0.0;
+  double angle = 0.0;
+  double val = 0.0;
   VALUE argument = GetRecord("angle", gCurrentContext);
 
   angle = TypeFloat(argument);
@@ -30,8 +30,8 @@ int MathSin(int arg_count){
 
 /*Math.cos(angle)*/
 int MathCos(int arg_count){
-  float angle = 0.0;
-  float val = 0.0;
+  double angle = 0.0;
+  double val = 0.0;
   VALUE argument = GetRecord("angle", gCurrentContext);
 
   angle = TypeFloat(argument);
@@ -45,8 +45,8 @@ int MathCos(int arg_count){
 
 /*Math.tan(angle)*/
 int MathTan(int arg_count){
-  float angle = 0.0;
-  float val = 0.0;
+  double angle = 0.0;
+  double val = 0.0;
   VALUE argument = GetRecord("angle", gCurrentContext);
 
   angle = TypeFloat(argument);
@@ -60,8 +60,8 @@ int MathTan(int arg_count){
 
 /*Math.asin(num)*/
 int MathAsin(int arg_count){
-  float input = 0.0;
-  float val = 0.0;
+  double input = 0.0;
+  double val = 0.0;
   VALUE argument = GetRecord("num", gCurrentContext);
 
   input = TypeFloat(argument);
@@ -75,8 +75,8 @@ int MathAsin(int arg_count){
 
 /*Math.acos(num)*/
 int MathAcos(int arg_count){
-  float input = 0.0;
-  float val = 0.0;
+  double input = 0.0;
+  double val = 0.0;
   VALUE argument = GetRecord("num", gCurrentContext);
 
   input = TypeFloat(argument);
@@ -90,8 +90,8 @@ int MathAcos(int arg_count){
 
 /*Math.atan(num)*/
 int MathAtan(int arg_count){
-  float input = 0.0;
-  float val = 0.0;
+  double input = 0.0;
+  double val = 0.0;
   VALUE argument = GetRecord("num", gCurrentContext);
 
   input = TypeFloat(argument);
@@ -105,8 +105,8 @@ int MathAtan(int arg_count){
 
 /*Math.exp(num)*/
 int MathExp(int arg_count){
-  float input = 0.0;
-  float val = 0.0;
+  double input = 0.0;
+  double val = 0.0;
   VALUE argument = GetRecord("num", gCurrentContext);
 
   input = TypeFloat(argument);
@@ -120,8 +120,8 @@ int MathExp(int arg_count){
 
 /*Math.log(num)*/
 int MathLog(int arg_count){
-  float input = 0.0;
-  float val = 0.0;
+  double input = 0.0;
+  double val = 0.0;
   VALUE argument = GetRecord("num", gCurrentContext);
 
   input = TypeFloat(argument);
@@ -135,8 +135,8 @@ int MathLog(int arg_count){
 
 /*Math.log10(num)*/
 int MathLog10(int arg_count){
-  float input = 0.0;
-  float val = 0.0;
+  double input = 0.0;
+  double val = 0.0;
   VALUE argument = GetRecord("num", gCurrentContext);
 
   input = TypeFloat(argument);
@@ -150,9 +150,9 @@ int MathLog10(int arg_count){
 
 /*Math.pow(base, exponent)*/
 int MathPow(int arg_count){
-  float base = 0.0;
-  float exponent = 0.0;
-  float val = 0.0;
+  double base = 0.0;
+  double exponent = 0.0;
+  double val = 0.0;
 
   VALUE base_arg = GetRecord("base", gCurrentContext);
   VALUE exp_arg = GetRecord("exponent", gCurrentContext);
@@ -170,8 +170,8 @@ int MathPow(int arg_count){
 
 /*Math.sqrt(num))*/
 int MathSqrt(int arg_count){
-  float input = 0.0;
-  float val = 0.0;
+  double input = 0.0;
+  double val = 0.0;
   VALUE argument = GetRecord("num", gCurrentContext);
 
   input = TypeFloat(argument);
@@ -186,8 +186,8 @@ int MathSqrt(int arg_count){
 
 /*Math.ceil(num)*/
 int MathCeil(int arg_count){
-  float input = 0.0;
-  float val = 0.0;
+  double input = 0.0;
+  double val = 0.0;
   VALUE argument = GetRecord("num", gCurrentContext);
 
   input = TypeFloat(argument);
@@ -202,8 +202,8 @@ int MathCeil(int arg_count){
 
 /*Math.floor(num)*/
 int MathFloor(int arg_count){
-  float input = 0.0;
-  float val = 0.0;
+  double input = 0.0;
+  double val = 0.0;
   VALUE argument = GetRecord("num", gCurrentContext);
 
   input = TypeFloat(argument);
@@ -218,8 +218,8 @@ int MathFloor(int arg_count){
 
 /*Math.abs(num)*/
 int MathAbs(int arg_count){
-  float input = 0.0;
-  float val = 0.0;
+  double input = 0.0;
+  double val = 0.0;
   VALUE argument = GetRecord("num", gCurrentContext);
 
   input = TypeFloat(argument);
